# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: > 
    The Mechanic Shop API allows clients to manage customers, mechanics, service tickets, inventory, and related operations.
    It provides RESTful endpoints to create, retrieve, update, and delete records, supporting functionality such as
    assigning mechanics to service tickets, updating inventory, assigning parts to service tickets, and managing customer profiles.
    This API is designed to streamline the workflow of a mechanic shop and integrate with frontend or mobile interfaces.
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
paths: 
...
    /service_tickets:
      # Get all service tickets
      get:
        tags:
          - "Service Tickets"
        summary: "Get all service tickets"
        description: "Retrieve a list of all service tickets."
        responses:
          200:
            description: "A list of service tickets."
            schema:
              type: array
              items:
                $ref: '#/definitions/ServiceTicket'
      
      # Create a new service ticket
      post:
        tags:
          - "Service Tickets"
        summary: "Create a new service ticket"
        description: "Add a new service ticket to the system."
        parameters:
          - in: body
            name: service_ticket
            required: true
            schema:
              $ref: '#/definitions/ServiceTicketInput'
        responses:
          201:
            description: "Service ticket created successfully."
            schema:
              $ref: '#/definitions/ServiceTicket'
      
    # Get, update, and delete a specific service ticket by ID
    /service_tickets/{ticket_id}:
      # Get a specific service ticket by ID
      get:
        tags:
          - "Service Tickets"
        summary: "Get a service ticket by ID"
        description: "Retrieve a specific service ticket by its ID."
        parameters:
          - name: ticket_id
            in: path
            required: true
            type: integer
        responses:
          200:
            description: "Service ticket details."
            schema:
              $ref: '#/definitions/ServiceTicket'
          404:
            description: "Service ticket not found."
    
      # Update a specific service ticket by ID
      put:
        tags:
          - "Service Tickets"
        summary: "Update a service ticket"
        description: "Update an existing service ticket's information, including assigned mechanics."
        parameters:
          - name: ticket_id
            in: path
            required: true
            type: integer
          - in: body
            name: service_ticket
            required: true
            schema:
              $ref: '#/definitions/ServiceTicketUpdateInput'
        responses:
          200:
            description: "Service ticket updated successfully."
            schema:
              $ref: '#/definitions/ServiceTicket'
          404:
            description: "Service ticket not found."
      
      # Delete a specific service ticket by ID
      delete:
        tags:
          - "Service Tickets"
        summary: "Delete a service ticket"
        description: "Remove a service ticket from the system."
        security:
          - BearerAuth: []
        parameters:
          - name: ticket_id
            in: path
            required: true
            type: integer
        responses:
          204:
            description: "Service ticket deleted successfully."
          404:
            description: "Service ticket not found."
    
    # Assign a product to a service ticket
    /service_tickets/{ticket_id}/add_product:
      post:
        tags:
          - "Service Tickets"
        summary: "Assign a product to a service ticket"
        description: "Assign a product with a quantity to a specific service ticket."
        parameters:
          - name: ticket_id
            in: path
            required: true
            type: integer
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                required:
                  - product_id
                  - quantity
                properties:
                  product_id:
                    type: integer
                    example: 1
                  quantity:
                    type: integer
                    example: 2
    
    # Get all service tickets for a specific mechanic or customer as authenticated user
    /service_tickets/my-tickets:
      get:
        tags:
          - "Service Tickets"
        summary: "Get all service tickets for current user"
        description: "Retrieve a list of all service tickets assigned to the current user."
        security:
          - BearerAuth: []
        responses:
          200:
            description: "A list of service tickets for the authorized user, wheter they are a mechanic or customer."
            schema:
              type: array
              items:
                $ref: '#/definitions/ServiceTicket'
          404:
            description: "Unauthorized."
          401:
            description: "Unauthorized. Token missing or invalid."

# ------------------------ Security Definitions -----------------------
# Security definitions for authentication
securityDefinitions:
    BearerAuth:
      type: "apiKey"
      name: "Authorization"
      in: "header"
      description: "Bearer token for authentication. You must log in via the /auth/login to receive a token. Example: 'Bearer your.token.here'"
definitions:
...
  ServiceTicketInput:
    type: object
    required:
      - customer_id
      - service_desc
    properties:
      customer_id:
        type: integer
        example: 1
      vin:
        type: string
        example: "1HGCM82633A123456"
      service_desc:
        type: string
        example: "Oil change and tire rotation"
      mechanic_ids:
        type: array
        items:
          type: integer
        example: [1, 2]
  ServiceTicket:
    type: object
    required:
      - id
      - customer_id
      - vin
      - service_desc
      - mechanic_ids
    properties:
      id:
        type: integer
        example: 1
      customer_id:
        type: integer
        example: 1
      service_desc:
        type: string
        example: "Oil change and tire rotation"
      mechanic_ids:
        type: array
        items:
          type: integer
        example: [1, 2]
  ServiceTicketUpdateInput:
      type: object
      required:
        - vin
        - service_desc
        - add_mechanic_ids
        - remove_mechanic_ids
      properties:
        vin:
          type: string
          example: "1HGCM82633A123456"
        service_desc:
          type: string
          example: "Oil change and tire rotation"
        add_mechanic_ids:
          type: array
          items:
            type: integer
          example: [1, 2]
        remove_mechanic_ids:
          type: array
          items:
            type: integer
          example: [1, 2]